changed exception handling
