+++ src/main/java/pan/artem/tinkoff/configuration/AppConfiguration.java
@@ -9,7 +9,7 @@ import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
-import pan.artem.tinkoff.controller.error.externalservice.ExternalServiceResponseErrorHandler;
+import pan.artem.tinkoff.controller.handler.ExternalServiceResponseErrorHandler;
 import pan.artem.tinkoff.properties.AppProperties;
 import pan.artem.tinkoff.service.CurrentWeatherService;
 import pan.artem.tinkoff.service.CurrentWeatherServiceImpl;
