+++ src/main/java/pan/artem/tinkoff/dto/ErrorInfo.java
@@ -1,4 +1,4 @@
-package pan.artem.tinkoff.controller.error;
+package pan.artem.tinkoff.dto;
 
 import lombok.Value;
