+++ src/main/java/pan/artem/tinkoff/exception/MyAppException.java
@@ -1,4 +1,4 @@
-package pan.artem.tinkoff.controller.error;
+package pan.artem.tinkoff.exception;
 
 public abstract class MyAppException extends RuntimeException {
