+++ src/main/java/pan/artem/tinkoff/exception/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package pan.artem.tinkoff.controller.error;
+package pan.artem.tinkoff.exception;
 
 public class ResourceNotFoundException extends MyAppException {
