+++ src/main/java/pan/artem/tinkoff/exception/externalservice/ExternalServiceInternalErrorException.java
@@ -1,4 +1,4 @@
-package pan.artem.tinkoff.controller.error.externalservice;
+package pan.artem.tinkoff.exception.externalservice;
 
 public class ExternalServiceInternalErrorException extends ExternalServiceException {
